pyproject.toml
[project]
name = "multi-agent-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "semantic-kernel>=1.36.0",
    "azure-ai-evaluation>=1.12.0",
]
